Update database schema to fix JobHighlightDAO bug
Each JobHighlight items List was previously stored in the db by storing
concatenating the List with "\n". Some items in the List contained line
breaks, causing a single item to be split into multiple items when a job
was retrieved from the db.

This was resolved by adding item to the primary key in JobHighlightDAO
and storing each entry in the items List as a separate record.